+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 9.0.6 (1/29/2024 - )
 
 * executor: tweak shutdown handling, print all tasks not complete
+* jre: published neowu/jre:21.0.2
 
 ### 9.0.5 (1/10/2024 - 1/29/2024)